src/requirements.txt
boto3==1.26.149
botocore==1.29.149
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
python-dotenv==0.21.0
